requirements.txt
Pillow~=8.3.2
reportlab~=3.6.1
PyQt5~=5.15.5